fix:예약글상세수정
